feat: Add getRemainingTime function to useTimer hook
